simplify launcher
